better dense fixed size tree

+++ grovedb-dense-fixed-sized-merkle-tree/src/store.rs
@@ -5,6 +5,6 @@ use crate::DenseMerkleError;
 /// Uses `&self` (interior mutability) to match the GroveDB `StorageContext`
 /// pattern.
 pub trait DenseTreeStore {
-    fn get_value(&self, position: u64) -> Result<Option<Vec<u8>>, DenseMerkleError>;
-    fn put_value(&self, position: u64, value: &[u8]) -> Result<(), DenseMerkleError>;
+    fn get_value(&self, position: u16) -> Result<Option<Vec<u8>>, DenseMerkleError>;
+    fn put_value(&self, position: u16, value: &[u8]) -> Result<(), DenseMerkleError>;
 }
