+++ grovedb-element/src/element/helpers.rs
@@ -207,7 +207,7 @@ impl Element {
             Element::CommitmentTree(_, count, ..) => Some(*count),
             Element::MmrTree(_, mmr_size, _) => Some(*mmr_size),
             Element::BulkAppendTree(_, count, ..) => Some(*count),
-            Element::DenseAppendOnlyFixedSizeTree(_, count, ..) => Some(*count),
+            Element::DenseAppendOnlyFixedSizeTree(_, count, ..) => Some(*count as u64),
             _ => None,
         }
     }
